Completely rewrote formatter
The original implementation had several problems when dealing with lists
('*'-lines).

Instead of basing the indenting for a line purely on the amount of
whitespace present in the original input file, the required indenting is
now decided by the program as it keeps track of the indenting a line
*should* have instead. This works far more reliably.

In addition, long lines are now also wrapped automatically.